requirements.txt
Django==1.7.1
djangorestframework==2.3.13


